feat: Add IPLD link-system store (#119)
Adds a Lens Store to the repo - storing lens configurations (and binaries) as IPLD blocks, allowing external systems (such as Defra) to implement P2P syncing of Lenses.

Also re-introduces transaction support, bumps to Go 1.24 (this was required for a dependency, I cant remember which), and introduces testo to lens.

The txn support and Go version bump are quite small, and I forgot about the Go bump, so I left them in this PR. If you really want I can pull them out, if the PR gets bogged down in review that might be sensible.